+++ Source/SwiftLintFramework/Rules/SingleTestClassRule.swift
@@ -9,23 +9,27 @@
 import Foundation
 import SourceKittenFramework
 
-public struct QuickSingleSpecRule: Rule, OptInRule, ConfigurationProviderRule {
+public struct SingleTestClassRule: Rule, OptInRule, ConfigurationProviderRule {
     public var configuration = SeverityConfiguration(.warning)
 
     public init() {}
 
     public static let description = RuleDescription(
-        identifier: "quick_single_spec",
-        name: "Quick Single Spec",
-        description: "Test files should contain a single QuickSpec class.",
+        identifier: "single_test_class",
+        name: "Single Test Class",
+        description: "Test files should contain a single QuickSpec or XCTestCase class.",
         kind: .style,
         nonTriggeringExamples: [
             "class FooTests {  }\n",
-            "class FooTests: QuickSpec {  }\n"
+            "class FooTests: QuickSpec {  }\n",
+            "class FooTests: XCTestCase {  }\n"
         ],
         triggeringExamples: [
             "↓class FooTests: QuickSpec {  }\n↓class BarTests: QuickSpec {  }\n",
-            "↓class FooTests: QuickSpec {  }\n↓class BarTests: QuickSpec {  }\n↓class TotoTests: QuickSpec {  }\n"
+            "↓class FooTests: QuickSpec {  }\n↓class BarTests: QuickSpec {  }\n↓class TotoTests: QuickSpec {  }\n",
+            "↓class FooTests: XCTestCase {  }\n↓class BarTests: XCTestCase {  }\n",
+            "↓class FooTests: XCTestCase {  }\n↓class BarTests: XCTestCase {  }\n↓class TotoTests: XCTestCase {  }\n",
+            "↓class FooTests: QuickSpec {  }\n↓class BarTests: XCTestCase {  }\n"
         ]
     )
 
@@ -40,7 +44,9 @@ public struct QuickSingleSpecRule: Rule, OptInRule, ConfigurationProviderRule {
     // MARK: - Private
 
     private func quickSpecs(in file: File) -> [[String: SourceKitRepresentable]] {
-        return file.structure.dictionary.substructure.filter { $0.inheritedTypes.contains("QuickSpec") }
+        return file.structure.dictionary.substructure.filter {
+            !$0.inheritedTypes.filter { testClasses().contains($0) }.isEmpty
+        }
     }
 
     private func toViolation(in file: File,
@@ -51,7 +57,11 @@ public struct QuickSingleSpecRule: Rule, OptInRule, ConfigurationProviderRule {
             return StyleViolation(ruleDescription: type(of: self).description,
                                   severity: configuration.severity,
                                   location: Location(file: file, byteOffset: offset),
-                                  reason: "\(numberOfSpecs) Quick Specs found in this file.")
+                                  reason: "\(numberOfSpecs) test classes found in this file.")
         }
+    }
+
+    private func testClasses() -> [String] {
+        return ["QuickSpec", "XCTestCase"]
     }
 }